Fix direction = "inherit" for offscreen canvas
Offscreen canvas was not inheriting the direction attribute in the
way the spec described (which I am in the process of reworking). When
the direction is inherit the offscreen should try to get it from a
canvas element it is associated with, or the document if there is no
associated element (i.e. it was not transferred).

Also add tests for canvas text direction = "inherit"

Bug: 390272618
Change-Id: I5f862a4e2337b94b2eaa721733725a860872a824